+++ src/main/java/ai/reveng/model/FunctionsDetailResponse.java
@@ -68,7 +68,7 @@ public class FunctionsDetailResponse {
   public static final String SERIALIZED_NAME_FUNCTION_VADDR = "function_vaddr";
   @SerializedName(SERIALIZED_NAME_FUNCTION_VADDR)
   @javax.annotation.Nonnull
-  private Integer functionVaddr;
+  private Long functionVaddr;
 
   public static final String SERIALIZED_NAME_FUNCTION_SIZE = "function_size";
   @SerializedName(SERIALIZED_NAME_FUNCTION_SIZE)
@@ -151,7 +151,7 @@ public class FunctionsDetailResponse {
   }
 
 
-  public FunctionsDetailResponse functionVaddr(@javax.annotation.Nonnull Integer functionVaddr) {
+  public FunctionsDetailResponse functionVaddr(@javax.annotation.Nonnull Long functionVaddr) {
     this.functionVaddr = functionVaddr;
     return this;
   }
@@ -161,11 +161,11 @@ public class FunctionsDetailResponse {
    * @return functionVaddr
    */
   @javax.annotation.Nonnull
-  public Integer getFunctionVaddr() {
+  public Long getFunctionVaddr() {
     return functionVaddr;
   }
 
-  public void setFunctionVaddr(@javax.annotation.Nonnull Integer functionVaddr) {
+  public void setFunctionVaddr(@javax.annotation.Nonnull Long functionVaddr) {
     this.functionVaddr = functionVaddr;
   }
