+++ src/main/java/ai/reveng/model/MatchedFunction.java
@@ -71,7 +71,7 @@ public class MatchedFunction {
   public static final String SERIALIZED_NAME_FUNCTION_VADDR = "function_vaddr";
   @SerializedName(SERIALIZED_NAME_FUNCTION_VADDR)
   @javax.annotation.Nonnull
-  private Integer functionVaddr;
+  private Long functionVaddr;
 
   public static final String SERIALIZED_NAME_MANGLED_NAME = "mangled_name";
   @SerializedName(SERIALIZED_NAME_MANGLED_NAME)
@@ -168,7 +168,7 @@ public class MatchedFunction {
   }
 
 
-  public MatchedFunction functionVaddr(@javax.annotation.Nonnull Integer functionVaddr) {
+  public MatchedFunction functionVaddr(@javax.annotation.Nonnull Long functionVaddr) {
     this.functionVaddr = functionVaddr;
     return this;
   }
@@ -178,11 +178,11 @@ public class MatchedFunction {
    * @return functionVaddr
    */
   @javax.annotation.Nonnull
-  public Integer getFunctionVaddr() {
+  public Long getFunctionVaddr() {
     return functionVaddr;
   }
 
-  public void setFunctionVaddr(@javax.annotation.Nonnull Integer functionVaddr) {
+  public void setFunctionVaddr(@javax.annotation.Nonnull Long functionVaddr) {
     this.functionVaddr = functionVaddr;
   }
