+++ src/main/java/ai/reveng/model/MatchedFunctionSuggestion.java
@@ -60,7 +60,7 @@ public class MatchedFunctionSuggestion {
   public static final String SERIALIZED_NAME_FUNCTION_VADDR = "function_vaddr";
   @SerializedName(SERIALIZED_NAME_FUNCTION_VADDR)
   @javax.annotation.Nonnull
-  private Integer functionVaddr;
+  private Long functionVaddr;
 
   public static final String SERIALIZED_NAME_SUGGESTED_NAME = "suggested_name";
   @SerializedName(SERIALIZED_NAME_SUGGESTED_NAME)
@@ -94,7 +94,7 @@ public class MatchedFunctionSuggestion {
   }
 
 
-  public MatchedFunctionSuggestion functionVaddr(@javax.annotation.Nonnull Integer functionVaddr) {
+  public MatchedFunctionSuggestion functionVaddr(@javax.annotation.Nonnull Long functionVaddr) {
     this.functionVaddr = functionVaddr;
     return this;
   }
@@ -104,11 +104,11 @@ public class MatchedFunctionSuggestion {
    * @return functionVaddr
    */
   @javax.annotation.Nonnull
-  public Integer getFunctionVaddr() {
+  public Long getFunctionVaddr() {
     return functionVaddr;
   }
 
-  public void setFunctionVaddr(@javax.annotation.Nonnull Integer functionVaddr) {
+  public void setFunctionVaddr(@javax.annotation.Nonnull Long functionVaddr) {
     this.functionVaddr = functionVaddr;
   }
